molecule: renamed subcommands and moved actions
