+++ completers/molecule_completer/cmd/converge.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/carapace-sh/carapace"
-	"github.com/carapace-sh/carapace-bin/completers/molecule_completer/cmd/action"
+	"github.com/carapace-sh/carapace-bin/pkg/actions/tools/molecule"
 	"github.com/spf13/cobra"
 )
 
@@ -19,7 +19,7 @@ func init() {
 	convergeCmd.Flags().StringP("scenario-name", "s", "default", "Name of the scenario to target")
 
 	carapace.Gen(convergeCmd).FlagCompletion(carapace.ActionMap{
-		"scenario-name": action.ActionScenarios(),
+		"scenario-name": molecule.ActionScenarios(),
 	})
 
 	rootCmd.AddCommand(convergeCmd)
