+++ completers/molecule_completer/cmd/destroy.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/carapace-sh/carapace"
-	"github.com/carapace-sh/carapace-bin/completers/molecule_completer/cmd/action"
+	"github.com/carapace-sh/carapace-bin/pkg/actions/tools/molecule"
 	"github.com/spf13/cobra"
 )
 
@@ -28,8 +28,8 @@ func init() {
 	destroyCmd.MarkFlagsMutuallyExclusive("parallel", "no-parallel")
 
 	carapace.Gen(destroyCmd).FlagCompletion(carapace.ActionMap{
-		"driver-name":   action.ActionDrivers(),
-		"scenario-name": action.ActionScenarios(),
+		"driver-name":   molecule.ActionDrivers(),
+		"scenario-name": molecule.ActionScenarios(),
 	})
 
 	rootCmd.AddCommand(destroyCmd)
