+++ completers/molecule_completer/cmd/login.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/carapace-sh/carapace"
-	"github.com/carapace-sh/carapace-bin/completers/molecule_completer/cmd/action"
+	"github.com/carapace-sh/carapace-bin/pkg/actions/tools/molecule"
 	"github.com/spf13/cobra"
 )
 
@@ -20,8 +20,8 @@ func init() {
 	loginCmd.Flags().StringP("scenario-name", "s", "default", "Name of the scenario to target")
 
 	carapace.Gen(loginCmd).FlagCompletion(carapace.ActionMap{
-		"host":          action.ActionInstances(loginCmd),
-		"scenario-name": action.ActionScenarios(),
+		// "host":          action.ActionInstances(loginCmd), // TODO
+		"scenario-name": molecule.ActionScenarios(),
 	})
 
 	rootCmd.AddCommand(loginCmd)
