+++ completers/molecule_completer/cmd/prepare.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/carapace-sh/carapace"
-	"github.com/carapace-sh/carapace-bin/completers/molecule_completer/cmd/action"
+	"github.com/carapace-sh/carapace-bin/pkg/actions/tools/molecule"
 	"github.com/spf13/cobra"
 )
 
@@ -24,8 +24,8 @@ func init() {
 	prepareCmd.MarkFlagsMutuallyExclusive("force", "no-force")
 
 	carapace.Gen(prepareCmd).FlagCompletion(carapace.ActionMap{
-		"driver-name":   action.ActionDrivers(),
-		"scenario-name": action.ActionScenarios(),
+		"driver-name":   molecule.ActionDrivers(),
+		"scenario-name": molecule.ActionScenarios(),
 	})
 
 	rootCmd.AddCommand(prepareCmd)
